scripts/SETUP_AUTOMATION.py
"""
Complete step-by-step guide to get browser automation working
"""

print("""
╔═══════════════════════════════════════════════════════════════════╗
║  🤖 ETIMAD BROWSER AUTOMATION - COMPLETE SETUP GUIDE              ║
╚═══════════════════════════════════════════════════════════════════╝

📋 STEP-BY-STEP INSTRUCTIONS:

━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

STEP 1: Configure Your Credentials
━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

Open browser_config.py and edit:

    USE_BROWSER_AUTOMATION = True
    ETIMAD_USERNAME = "your_username_or_email"
    ETIMAD_PASSWORD = "your_password"
    HEADLESS_BROWSER = False  # Keep False to see what happens

━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

STEP 2: Test the Automation
━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

Run:
    python test_automation.py

What will happen:
✅ Chrome opens automatically
✅ Goes to Etimad website
✅ Clicks "تسجيل الدخول"
✅ Enters your credentials
✅ Logs in automatically
✅ Extracts cookies
✅ Tests cookies with API
✅ Saves to cookies_backup.json

Expected output:
    🤖 TESTING FULL BROWSER AUTOMATION
    ✅ SUCCESS!
    ✅ API Test Passed! 259 tenders available
    🎉 AUTOMATION IS FULLY WORKING!

━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

STEP 3: Run Your App
━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

Now just run:
    python app.py

The app will:
1. Check if cookies_backup.json exists (fast!)
2. If not, run browser automation automatically
3. Use cookies for all API calls
4. No manual copying needed! 🎉

Open: http://localhost:5000

━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

🚨 TROUBLESHOOTING
━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

Problem: "Login failed - check credentials"
Solution:
  → Verify username/password in browser_config.py
  → Try logging in manually first
  → Check for CAPTCHA or 2FA

Problem: "ChromeDriver not found"
Solution:
  → webdriver-manager should auto-download it
  → Try: pip install --upgrade webdriver-manager

Problem: "Element not clickable"
Solution:
  → Already handled with JavaScript click!
  → If still fails, check inspect_login_page.py output

Problem: "Cookies not working with API"
Solution:
  → Check if cookies contain 'MobileAuthCookie'
  → Try increasing wait time after login
  → Verify you're logging into the correct account type

━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

📚 AVAILABLE SCRIPTS
━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

inspect_login_page.py
  → Inspect Etimad login page
  → Finds correct selectors
  → Saves HTML files for analysis

test_automation.py
  → Test full automation end-to-end
  → Verify login works
  → Test API with cookies

test_browser_cookies.py
  → Alternative: Extract cookies from your logged-in browser
  → No automation needed
  → Must login manually first

app.py
  → Main Flask application
  → Smart cookie management
  → Auto-runs automation if needed

━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

🔒 SECURITY NOTES
━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

IMPORTANT: Don't commit credentials to Git!

Add to .gitignore:
    browser_config.py
    cookies_backup.json
    .env

For production, use environment variables:
    1. Create .env file:
       ETIMAD_USERNAME=your_username
       ETIMAD_PASSWORD=your_password
    
    2. Update browser_config.py:
       import os
       from dotenv import load_dotenv
       load_dotenv()
       ETIMAD_USERNAME = os.getenv('ETIMAD_USERNAME')
       ETIMAD_PASSWORD = os.getenv('ETIMAD_PASSWORD')

━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

💡 SMART COOKIE MANAGEMENT
━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

The app uses a smart 3-tier approach:

1. 📁 File Cache (cookies_backup.json)
   ✅ Fastest (<1ms)
   ✅ Reused until expiry
   
2. 🤖 Browser Automation
   ✅ Automatic fresh login
   ✅ Saves to file for next time
   
3. 📋 Manual Config (config.py)
   ✅ Fallback if everything fails

This means:
→ First run: Automation logs in (~10 seconds)
→ Next runs: Uses cached cookies (<1 second)
→ When expired: Auto re-authenticates

No manual copying ever needed! 🎉

━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

📖 MORE INFORMATION
━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

See detailed guides in docs/ folder:
  → BROWSER_AUTOMATION_GUIDE.md
  → COOKIE_MANAGEMENT_GUIDE.md

━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

Ready to start? Run:
    python test_automation.py

╔═══════════════════════════════════════════════════════════════════╗
║  Good luck! 🚀                                                     ║
╚═══════════════════════════════════════════════════════════════════╝
""")
